% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allign_ts.R
\name{allign_ts}
\alias{allign_ts}
\title{Aligning Time-Series length}
\usage{
allign_ts()
}
\value{
a list of time series starting with "ts_" followed by the column name followed
by "_adj", but now each time-series should have the same length.
}
\description{
This function takes no argument, and will align your corrected time series that
may have different length, because of differentiations.
}
\examples{
allign_ts()

}
