% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atypical_tso.R
\name{atypical_tso}
\alias{atypical_tso}
\title{Correct Time-Series from atypical points}
\usage{
atypical_tso()
}
\value{
a list of time series starting with "ts_" followed by the column name followed by "_adj"
}
\description{
This function takes no argument, and will correct your previously created time-series
and correct any atypical points. This function use the tso() function from the package tsoutliers
}
\examples{
atypical_tso()

}
