% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_models.R
\name{eco_models}
\alias{eco_models}
\title{Forecasting with 6 econometrics models}
\usage{
eco_models(DATAFRAME, Y_VARIABLE, PERIOD)
}
\arguments{
\item{DATAFRAME}{an R dataframe}

\item{PERIOD}{a frequency value, e.g. 12 for monthly data, 4 for quarterly data, etc.}

\item{Y_VARIALBE}{the Yt variable you're trying to predict}
}
\value{
RMSE for each models and the predictions
}
\description{
This function takes three arguments: DATAFRAME: Your dataframes with \( n \)
rows and \( p \) columns (with \( n > p \)). Y_VARIABLE: Your \( Y_t \) variable,
meaning the variable you're trying to predict. PERIOD: The frequency of your
time-series (12 for monthly, 4 for quarterly, etc...). This function uses the
following models and packages: ARX model with gets (gets). ARX model with auto.arima
(forecast). ARMAX model with auto.arima (forecast). Naive model (forecast).
LM model (stats). AR1 model (forecast). GAM model (mgcv).A short description...
}
\examples{
eco_models(df, "my_Y_variable", 12)
eco_models(result_df, "Food", 4)
eco_models(datafrale, "VIX", 12)

}
