% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_models.R
\name{ml_models}
\alias{ml_models}
\title{Forecasting with 6 Machine-Learning models}
\usage{
ml_models(DATAFRAME, Y_VARIABLE)
}
\arguments{
\item{DATAFRAME}{an R dataframe}

\item{Y_VARIALBE}{the Yt variable you're trying to predict}
}
\value{
RMSE for each models and the predictions
}
\description{
Machine-Learning Forecasting Models This function takes three arguments:
DATAFRAME: Your dataframes with \( n \) rows and \( p \) columns (with \( n > p \)).
Y_VARIABLE: Your \( Y_t \) variable, meaning the variable you're trying to predict.
This function uses the following models and packages: MLP models (neuralnet).
MARS model (earth). SVM model (e1071). Random Forest (randomForest).
XGB Boost (xgboost). kNN (caret).
}
\examples{
ml_models(df, "my_Y_variable")
ml_models(result_df, "Food")
ml_models(datafrale, "VIX")

}
