% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaso_correct.R
\name{seaso_correct}
\alias{seaso_correct}
\title{Seasonality Correction}
\usage{
seaso_correct()
}
\value{
a list of time series starting with "ts_" followed by the column name followed by "_adj"
}
\description{
This function takes no argument, and will correct any seasonality on your previously
created time-series. This function use the stl() function from the package stats.
}
\examples{
seaso_correct()

}
