% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaso_detect.R
\name{seaso_detect}
\alias{seaso_detect}
\title{Seasonality Detection}
\usage{
seaso_detect()
}
\value{
a list of time series starting with "ts_" followed by the column name followed by "_adj"
}
\description{
This function takes no argument, and will detect any seasonality on your previously
created time-series. This function use the combined_test() and seasdum() functions
from the package seastests.
}
\examples{
seaso_detect()

}
