% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaso_verif.R
\name{seaso_verif}
\alias{seaso_verif}
\title{Seasonality Verification}
\usage{
seaso_verif()
}
\value{
a list of time series starting with "ts_" followed by the column name followed by "_adj"
}
\description{
This function takes no argument, and will detect any seasonality on your previously
created and corrected time-series. This function use the combined_test() and seasdum()
functions from the package seastests. The main purpose of this function is to
ensure that there is no seasonality in the corrected series.
}
\examples{
seaso_verif()

}
