% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statio.R
\name{statio}
\alias{statio}
\title{Stationarity Treatment}
\usage{
statio()
}
\value{
a list of time series starting with "ts_" followed by the column name followed by "_adj"
}
\description{
This function takes no argument, and will detect and correct, if necessary,
the time series that are not stationary. The function detect stationarity with
the adf.test() function from the package tseries.
}
\examples{
statio()

}
