% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_transfo.R
\name{ts_transfo}
\alias{ts_transfo}
\title{Transform dataframe columns into Time Series}
\usage{
ts_transfo(DATAFRAME, YEAR, MONTH, FREQUENCY)
}
\arguments{
\item{DATAFRAME}{an R dataframe}

\item{YEAR}{a Year value, e.g. 2019}

\item{MONTH}{a Month value, e.g. 05 for May}

\item{FREQUENCY}{a frequency value, e.g. 12 for monthly data, 4 for quarterly data, etc.}
}
\value{
a list of time series starting with "ts_" followed by the column name
}
\description{
Transform dataframe columns into Time Series
}
\examples{
ts_transfo(df, 2019, 05, 12)
ts_transfo(df, 1994, 06, 4)

}
