% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vselec.R
\name{vselec}
\alias{vselec}
\title{Merge time-series into a single dataframe}
\usage{
vselec(Y_VARIABLE)
}
\arguments{
\item{first_column}{the name of the Y-variable column that you have to put first in the dataframe}
}
\value{
a data frame with the Y-variable (the one that your trying to predct) as the first column.
}
\description{
Merge time-series into a single dataframe
}
\examples{
create_df(df, "my_Y_variable")
create_df(dataframe, "Gas_Price")
create_df(financial_data, "VIX")

}
